Add files via upload
This R code contains a helper function that creates marginal and conditional R2 (r-squared) values for nlme and lme4 models. The function also creates a labels data.frame for easy plotting into ggplot2, in addition to a list of labels that can be used in R base graphics. Any help in improving the code or adding functionality (multilevel Rc values for lme4 for instance) would be much appreciated.